+++ tensorflow_datasets/public_api.py
@@ -34,9 +34,9 @@ from tensorflow_datasets.core.dataset_utils import as_numpy
 from tensorflow_datasets.core.download import GenerateMode
 from tensorflow_datasets.core.load import builder
 from tensorflow_datasets.core.load import builder_cls
-from tensorflow_datasets.core.load import builder_from_directory
 from tensorflow_datasets.core.load import list_builders
 from tensorflow_datasets.core.load import load
+from tensorflow_datasets.core.load_from_directory import builder_from_directory
 from tensorflow_datasets.core.splits import even_splits
 from tensorflow_datasets.core.splits import Split
 from tensorflow_datasets.core.utils.gcs_utils import is_dataset_on_gcs
